requirements.txt
Django
Pillow
djangorestframework
markdown
django-filter
